fix: replace test-d.ts file with normal types.spec.ts
