+++ packages/react-storage/jest.config.ts
@@ -4,8 +4,6 @@ const config: Config = {
   collectCoverage: true,
   collectCoverageFrom: [
     '<rootDir>/src/**/*.(ts|tsx)',
-    // do not collect from type definition files
-    '<rootDir>/src/**/*.test-d.ts',
     // do not collect from index, testUtils or version files
     '!<rootDir>/**/(index|version).(ts|tsx)',
     // do not collect from top level styles directory
@@ -29,7 +27,6 @@ const config: Config = {
   testPathIgnorePatterns: [
     '<rootDir>/src/components/StorageBrowser/displayText/libraries/en/__tests__/scenarios.ts',
     '__testUtils__/',
-    '.*\\.test-d\\.ts$',
   ],
   preset: 'ts-jest',
   setupFilesAfterEnv: ['./jest.setup.ts'],
